In the write_slab function, use WPSname from arg list rather than v.WPSname
Rather than using (probably unintentionally) v.WPSname when writing a slab of
data with a call to intfile.write_next_met_field in the write_slab function,
use the value of WPSname from the argument list instead. This is necessary to
permit calls to write_slab for fields that are different from the current value
of v.WPSname.