Implement diagnostic computation for several fields: SNOW, SNOWH, and RH at sfc
This commit introduces a means of providing an arbitrary set of diagnostic field
calculations. Demonstrating the approach are the computation of three surface
fields:

 - SNOW (water equivalent snow depth) from ERA5 RSN
 - SNOWH (physical snow depth) from SNOW and ERA5 SD
 - RH (relative humidity in %) from ERA5 VAR_2T and VAR_2D

In general, a new diagnostic field can be implemented as class with a 'consider'
method, whose job is to save information needed to compute the diagnostic and
when sufficient information is available, to compute and write the diagnostic
field. The diagnostic computation is registered by adding an instance of its
class to a list. The 'consider' method for each diagnostic is invoked for every
field at every level that is processed from ERA5 netCDF files.