Permit specific fields to be omitted from output intermediate file
Some fields (e.g., GEOPT and DEWPT) are only read from ERA5 netCDF files so that
they may be used in diagnosing other fields (e.g., GHT and RH). So that these
fields are not unnecessarily written to the output intermediate file, this
commit adds a new list, dont_output, to the main function. Fields whose WPS name
appears in the dont_output list will not be written to the intermediate file,
but they are still provided to all diagnostic functions.