If an ERA5 netCDF file is not found, print a useful error message
The find_era5_file function now raises an exception if the ERA5 netCDF file
needed for a given variable is not found. The main function catches this
exception and prints an error message indicating the filename that could not
be found as well as the name of the ERA5 variable.