Avoid double-negatives in logic for writing output fields
Previously, what could be viewed as a double-negative was present in the logic
for determining if a field should be written to the output intermediate file.
In an attempt to improve readability, this commit changes the logic so that we
test whether a variable is in the "don't output" list rather than testing
whether a variable is *not* in the "don't output" list.